fix(plugins): update duration field to use seconds as float instead of milliseconds as uint32

+++ plugins/capabilities/metadata_agent.go
@@ -150,8 +150,8 @@ type SongRef struct {
 	Album string `json:"album,omitempty"`
 	// AlbumMBID is the MusicBrainz release ID.
 	AlbumMBID string `json:"albumMbid,omitempty"`
-	// Duration is the song duration in milliseconds.
-	Duration uint32 `json:"duration,omitempty"`
+	// Duration is the song duration in seconds.
+	Duration float32 `json:"duration,omitempty"`
 }
 
 // TopSongsResponse is the response for GetArtistTopSongs.
